tests: increase timeout in  getnettotals test
Summary:
We saw test failure in a CI run, because the time between the first `ping` and the last `pong` was about 4 seconds and the response for `getnettotals` was not up to date after one second.
Bumping  the timeouts from 1 s to 10 s should allow the nodes to synchronize their states and make the test more reliable when it is run on a slow system.

Test Plan: `ninja check-functional`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10723