[secp256k1] tests: Fix test whose result is implementation-defined
Summary:
A compiler may add struct padding and fe_cmov is not guaranteed to
preserve it.

On the way, we improve the identity check such that it covers the
VERIFY struct members.

This is a backport of [[https://github.com/bitcoin-core/secp256k1/pull/1054 | secp256k1#1054]]

Test Plan: `ninja check-secp256k1`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D19473