[secp256k1] Initial gej blinding -> final ge blinding
Summary:
Instead of having the starting point of the ecmult_gen computation be
offset, do it with the final point. This enables reasoning over the
set of points reachable in intermediary computations, which can be
leveraged by potential future optimization.

Because the final point is in affine coordinates, its projective
blinding is no longer possible. It will be reintroduced again in
a different way, in a later commit.

Also introduce some more comments and more descriptive names.

This is a partial backport of [[https://github.com/bitcoin-core/secp256k1/pull/1058 | secp256k1#1058]]
https://github.com/bitcoin-core/secp256k1/pull/1058/changes/ab45c3e0892ec6cffed714d89fb146bee694fcb2

Depends on D20234

Test Plan: `ninja check-secp256k1`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D20235