[avalanche] Update our position according to the vote
Summary:
Accept or reject a proof being voted on according to the avalanche consensus.
There is still a missing cooldown feature (prevent proof replacement after a vote has been accepted and finalized, see D10714 summary) which will be added in a follow up. This is fine because the feature is disabled by default by the replacement option flag.

Ref T1854.

Depends on D10727.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, tyler-smith

Reviewed By: #bitcoin_abc, tyler-smith

Subscribers: tyler-smith

Maniphest Tasks: T1854

Differential Revision: https://reviews.bitcoinabc.org/D10788