added timezone constant for all npm tests
Summary:
Allowing npm tests to run in UTC timezone.
This is needed because we have a jest test that checks a DateTime formatter function.
However it formats the date time to the local machines timezone, which causes the test to fail.
This change fixes that failing test no matter which time zone the tests are run.

Test Plan:
Change your machines to different time zones and run the npm tests.

```bash
cd bitcoin-abc/web/cashtab
npm i
npm test
# all tests should pass!
```

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, bytesofman

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, bytesofman

Differential Revision: https://reviews.bitcoinabc.org/D10809