Unbreak the build with boost 1.78
Summary:
The recent fs.h changes broke the build with boost 1.78 (see https://github.com/bitcoin/bitcoin/issues/23846), but the piece of code causing the issue is currently unused in our codebase. This diff comments it out until a fix is merged that can be backported (expected to happen soon).

Note: boost 1.78 introduces another issue (change of behavior for boost::filsystem::canonical) which can be worked around by disabling the wallet, so I leave it for now. This will be fixed at the same time as this issue.

Test Plan:
With boost 1.78:
  ninja all check

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D10822