Add an option to override the max addr sent in response to a getaddr message
Summary:
This will make it possible to write tests for the avalanche node addresses request and responses. Since this is intended for testing only this is put in the hidden options, there is no boundary check and the default causes no change in behavior (tested in p2p_getaddr_caching.py).

Ref T1696.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Maniphest Tasks: T1696

Differential Revision: https://reviews.bitcoinabc.org/D10823