[avalanche] Add a getavaaddr message
Summary:
This is part of the active peer discovery. This new message is equivalent to the standard getaddr message and asks for the best avalanche peers, i.e. the ones with the highest availability score. The response use the standard addr message.

For now this does not come with any of the privacy features of the getaddr message, this can be added in follow-up diffs if needed. The messages are limited to one every 10 minutes to prevent a peer from exhausting our resources by spamming with getavaaddr.

Ref T1696.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, tyler-smith

Reviewed By: #bitcoin_abc, tyler-smith

Subscribers: tyler-smith

Maniphest Tasks: T1696

Differential Revision: https://reviews.bitcoinabc.org/D10830