Move SetupSeederArgs out of parsing code to make it easier to test
Summary:
Trying to add more test cases to `options_tests.cpp` results in errors
due to the way the seeder sets up the ArgsManager on each call to ParseCommandLine.
It doesn't make sense to setup args while parsing anyway, so moving it feels right.

The patch includes adding a test case to demonstrate that the change works as intended.

Test Plan:
  ninja check-seeder bitcoin-seeder

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, Fabien

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10828