Replace `struct update_lock_points` with lambda
Summary:
No behavior change.
This code was introduced in 5add7a7 before we required C++11, which is
why the struct was needed. As we are now using more modern C++ and this
is the only place where lockpoints are updated for mempool entries, it
is more idiomatic to call `modify` with a lambda.

Co-authored-by: Hennadii Stepanov <32963518+hebasto@users.noreply.github.com>

This concludes backport of [[https://github.com/bitcoin/bitcoin/pull/23976 | core#23976]]
https://github.com/bitcoin/bitcoin/pull/23976/commits/e177fcab3831b6d259da5164cabedcc9e78f6957

Depends on D12587

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D12588