[Cashtab] Add custom styling to Antd Descriptions
Summary: In order to make further progress with the dark theme, the antd Descriptions in SendToken.js need custom CSS. Currently, it is not possible to change the background and foreground colors in left most column on the table that contains the token's stats. If left unchanged the column would maintain its white background and dark text when the app is toggled to the dark theme. In order to exercise more control over the table a whole, I have included both columns (.ant-descriptions-item-lable & .ant-descriptions-item-content) in AntDescriptionsCss.  Related to task T2173.

Test Plan:
cd web/cashtab
npm start
navigate to send token page
ensure that content in the table is legible
In your IDE:
In theme.js, change the value of props.theme.tokenListItem.background and props.theme.tokenListItem.color
In browser:
navigate to the Send Token page.
Ensure that the colors selected in the previous step match what is rendering on the screen.

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D10839