doc, test: Improve setup_clean_chain documentation
Summary:
```
The first commit improves documentation on setup_clean_chain which is misunderstood quite frequently. Most importantly it fixes the TestShell docs which are simply incorrect.

The second commit removes the instances of setup_clean_clain in functional tests where it is not changing the default.
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/21042 | core#21042]].

Test Plan:
  ninja check-functional-extended

In `test/functional`, run:
  git grep 'self.setup_clean_chain = False'
and check the is no occurrence in the test outside of the framework itself.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D10853