Increase the timeout for sending a large block in abc_p2p_fullblocktest
Summary:
This test fails sometimes due to a timeout on the TSAN or OSX builds.
The timeout has already been bumped for RPC submission, so do the same for p2p submission. Worst case the test takes longer to fail (if there is a deadlock) which is annoying, but will clearly indicate the source is not resource exhaustion. In the best case this fixes the flakyness.

Test Plan:
With TSAN:
  ./test/functional/test_runner.py abc_p2p_fullblocktest

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D10854