tests: Test that a fully signed tx given to signrawtx is unchanged
Summary:
> Tests that a fully signed transaction given to
> `signrawtransactionwithwallet` is both unchanged and marked as complete.
> This tests for a regression in Bitcoin Core 0.20 where the transaction would not be
> marked as complete.

We do not have the regression in Bitcoin ABC and it seems unlikely that it will happen, as it was bissected to [[https://github.com/bitcoin/bitcoin/pull/16528 | core#16528]] which we already finished backporting, and had to be fixed only on Core 0.20 (not master). But just in case, it is better to have the test to avoid any bad surprise.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/20562 | core#20562]]

Test Plan: `ninja check-functional`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10894