Remove utxo db upgrade code
Summary:
We already require a full reindex for pre-0.22.8 versions (see D11355) so this upgrade code path from pre-0.15.0 versions is pointless and can be removed.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/24236 | core#24236]]

Test Plan:
This removes an untested code path, which has the lucky side-effect of increasing relative test coverage. So we just check that no existing test is accidentaly broken..
`ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18972