Bump the timeout for the debug log  messages in test_seed_peers
Summary: There have been a few CI failures on this test. From investigation the failures occurs on assertions that are expecting several debug log messages to appear upon node startup. The default timeout is used which is 2s and sometimes this is not enough for all the messages to be printed. This diff bumps the timeout for the cases where several messages are expected after a node restart to get rid of the failures.

Test Plan:
  ./test/functional/test_runner.py feature_config_args

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D10984