[avalanche] Extend our outbound connections with avalanche enabled peers
Summary:
This adds a new AVALANCHE_OUTBOUND connection type which is just a special case of a FULL_OUTBOUND_RELAY peer which also provides the avalanche service. We add an extra 16 outbound slots reserved for these peers, which will be the ones we will be sending the getavaaddr messages to. If avalanche is not enabled these slots are not created and there is no change in behavior. The eviction and replacement mechanism is kept identical to the usual full outbound peers.

Ref T1696.

Depends on D11009.

Test Plan:
  ninja check-extended

Reviewers: #bitcoin_abc, tyler-smith

Reviewed By: #bitcoin_abc, tyler-smith

Subscribers: tyler-smith

Maniphest Tasks: T1696

Differential Revision: https://reviews.bitcoinabc.org/D11010