[avalanche] Detect when more nodes are needed
Summary:
If the peer manager fails to find an available node during a call to `selectNode()`, it will latch a flag that can be used to trigger requesting more avalanche node addresses. The latch is atomic and cleared on read so we can loop over it in a network thread (out of scope for this diff).

Ref T1696.

Test Plan:
  ninja check-avalanche

Reviewers: #bitcoin_abc, tyler-smith

Reviewed By: #bitcoin_abc, tyler-smith

Subscribers: tyler-smith

Maniphest Tasks: T1696

Differential Revision: https://reviews.bitcoinabc.org/D11034