[Cashtab] Add support for Swiss Franc
Summary: Added support for Swiss Franc (CHF). One issue is, if using the internationally accepted "CHF" for the symbol, the amount appears as "CHF [amount] CHF". According to some sources, Fr. is an acceptable currency symbol, which displays the amount as "Fr. [amount] CHF" which I think is easier on the eyes. That said, it can be changed back if that is preferrable. Related to task T2210.

Test Plan:
cd web/cashtab && npm start
navigate to settings tab
select Swiss Franc from dropdown menu
navigate to wallet tab and check that CHF is the selected currency
using an online conversion tool, ensure conversion is taking place

Reviewers: bytesofman, emack, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D11038