[Cashtab] [utxo optimization part 1] Add function isExcludedUtxo
Summary:
T1683

See D11039 for a working alpha of the end goal of this series of diffs.

X Add function isExcludedUtxo with unit tests (determines if a given utxo is in an array of utxos)
_ Add function whichUtxosWereAdded with unit tests
_ Add function whichUtxosWereRemoved with unit tests
_ Add function addNewHydratedUtxos with unit tests
_ Add function removeConsumedUtxos with unit tests
_ Add function validateIncrementedHydratedUtxoDetails with unit tests
_ Using the above functions, modify the `update` function in `useBCH.js` to only hydrate (that is, collect eToken info about) the utxos that changed since the last time `update` ran
_ Reduce the interval between calls to `update`, speeding up Cashtab refresh rate

Current state of Cashtab:
`hydratedUtxoDetails` is recreated each time the utxo set changes. This requires a series of slow API calls to SLPDB.

Cashtab after this series of diffs lands:
Only new incoming utxos are parsed by API calls to SLPDB. These are added to `hydratedUtxoDetails`. Consumed utxos are removed. The resulting `hydratedUtxoDetails` is then validated to ensure it includes all utxos in the wallet. In testing, this reduces Cashtab UI locktime after sending a tx by about 60%.

Test Plan:
1. Review the new function `isExcludedUtxo`
2. Review the unit tests. Confirm no others are required, or suggest some others.
3. Note: this new function is not used anywhere in Cashtab. This diff does not alter the behavior of the app.

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D11045