[LINTER] Get rid of cppcheck
Summary:
Cppcheck has been introduced as linter in order to find the missing `explicit` statements in single parameter constructors.
However this has been very inefficient for various reasons:
 - There are more exceptions in the linter than it actually found issues
 - Cppcheck is broken every couple releases: 2.4 was OK, 2.5 was buggy, 2.6 was OK, and now 2.7 is also buggy.

As a consequence we spent more time working around the linter than actually using it to detect issues, which is a net negative.

Test Plan:
  arc lint --everything

With Clang and Debug (sanity check for peermanager.cpp change):
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D11063