coinstats: Move hasher codepath to kernel/coinstats
Summary:
As mentioned in a previous commit, the hashing codepath can now be moved
to a separate file. This decouples callers that only rely on the hashing
codepath from the indexing one.

This is key for libbitcoinkernel, which needs to have the CoinsStats
hashing codepath for AssumeUTXO, but does not wish to be coupled with
indexes.

Note that only the .cpp file is split in this commit, the header files
will be split in a subsequent commit and the #includes to
node/coinstats.h will be adjusted to only #include the necessary
headers.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/24410 | core#24410]]
https://github.com/bitcoin/bitcoin/pull/24410/commits/35f73ce4b2efd7341fe55f77b334f27ad8aad090

Depends on D16097

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16098