Output license info when binaries are passed -version
Summary:
Consolidate to outputting the licensing info when we pass -version to a binary.
Prior to this commit, the behavior is inconsistent between the various binaries.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/24409 | core#24409]]
https://github.com/bitcoin/bitcoin/pull/24409/commits/2618fb8d15d01dca967856c92ebf3e4cc09699a2

Depends on D16215

Test Plan:
```
ninja check-functional
src/bitcoind -version
src/bitcoin-cli -version
src/bitcoin-wallet -version
src/bitcoin-tx -version

```
Test that the output of the manpage generation script makes sense for  bitcoin-tx and bitcoin-wallet

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16216