[Cashtab] [utxo opt. part 4 of 8] Add function addNewHydratedUtxos
Summary:
T1683

See D11039 for a working alpha of the end goal of this series of diffs.

Note: After review, I'm not sure we really need the isValidUtxo function introduced in D11045. All of these functions are comparing API responses from the same endpoint a few seconds apart from each other. We do not have any experience of bch-api occasionally returning a valid utxos response with the exception of some utxos missing parameters.

However, I'll leave it there until we get to the end of this series of diffs -- then pull it out if it is not used.

X Add function isExcludedUtxo with unit tests (determines if a given utxo is in an array of utxos)
X Add function whichUtxosWereAdded with unit tests
X Add function whichUtxosWereRemoved with unit tests
X Add function addNewHydratedUtxos with unit tests
_ Add function removeConsumedUtxos with unit tests
_ Add function validateIncrementedHydratedUtxoDetails with unit tests
_ Using the above functions, modify the update function in useBCH.js to only hydrate (that is, collect eToken info about) the utxos that changed since the last time update ran
_ Reduce the interval between calls to update, speeding up Cashtab refresh rate

Current state of Cashtab:
hydratedUtxoDetails is recreated each time the utxo set changes. This requires a series of slow API calls to SLPDB.

Cashtab after this series of diffs lands:
Only new incoming utxos are parsed by API calls to SLPDB. These are added to hydratedUtxoDetails. Consumed utxos are removed. The resulting hydratedUtxoDetails is then validated to ensure it includes all utxos in the wallet. In testing, this reduces Cashtab UI locktime after sending a tx by about 60%.

Test Plan:
Review the new function addNewHydratedUtxos. Note: this function is specific to the data structures returned by bch-api and is not expected to work with invalid data.
Review the unit tests. Confirm no others are required, or suggest some others.
Note: this new function is not used anywhere in Cashtab. This diff does not alter the behavior of the app.

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D11088